update all service code

+++ src/services/cart_service.py
@@ -1,3 +1,5 @@
+from werkzeug import exceptions
+
 from src.models import Cart
 from src.services import database_service
 
@@ -7,6 +9,10 @@ def create():
 
 
 def delete(cart_id):
+    cart = get(cart_id)
+    if cart['is_ordered']:
+        raise exceptions.BadRequest(f'You cannot delete cart with id "{cart_id}" that has an associated order.')
+
     return database_service.delete_entity_instance(Cart, cart_id)
